move user components to user folder
